@@ -15,28 +27,25 @@
     "build"
   ],
   "scripts": {
-    "build": "rollup -c",
-    "storybook": "start-storybook -p 8008 --no-open",
-    "storybook:build": "build-storybook --quiet",
+    "build": "vite build",
+    "storybook": "cross-env STORYBOOK=true TS_NODE_PROJECT=.storybook/tsconfig.json start-storybook -p 8008",
+    "storybook:build": "cross-env STORYBOOK=true TS_NODE_PROJECT=.storybook/tsconfig.json build-storybook --quiet",
     "test:storybook:visual": "./visual-regressions/bin/take-and-check-in-docker.sh",
     "test:storybook:visual:update": "./visual-regressions/bin/take-and-update-in-docker.sh",
     "test:storybook:visual:ci": "./visual-regressions/bin/take-and-check-in-docker-ci.sh",
-    "test": "jest -w 1",
+    "test": "jest -w 4",
     "test:watch": "jest --watch",
-    "types": "tsc --noEmit",
+    "types": "tsc -b",
     "lint": "eslint 'src/**/*.{js,ts,tsx}'",
     "lint:fix": "eslint 'src/**/*.{js,ts,tsx}' --fix",
     "lint:css": "stylelint --config .stylelintrc-css.json 'src/**/*.css'",
     "lint:css-jsx": "stylelint --config .stylelintrc-jsx.json 'src/**/*.{ts,tsx}'",
     "semantic-release": "semantic-release",
-    "betterer": "betterer -t ./tsconfig.json -c .betterer/.betterer.ts -r .betterer/.betterer.results --strict",
-    "betterer:precommit": "betterer precommit -t ./tsconfig.json -c .betterer/.betterer.ts -r .betterer/.betterer.results --strict",
-    "betterer:ci": "betterer ci -t ./tsconfig.json -c .betterer/.betterer.ts -r .betterer/.betterer.results --strict",
     "prepare": "husky install"
   },
   "peerDependencies": {
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^5.2.0",
     "react-select": "^4.3.1",
     "styled-components": "^5.1.0"
@@ -149,22 +161,15 @@
     "postcss-styled-syntax": "^0.6.4",
     "postcss-syntax": ">=0.36.2",
     "prettier": "^2.6.0",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-is": ">= 16.8.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-is": ">= 18.3.1",
     "react-router-dom": "^5.3.0",
     "react-select": "^4.3.1",
     "react-select-event": "^5.3.0",
-    "react-test-renderer": "^16.13.1",
-    "require-from-string": "^2.0.2",
-    "rollup": "^2.70.1",
     "rollup-plugin-auto-external": "^2.0.0",
     "rollup-plugin-delete": "^2.0.0",
-    "rollup-plugin-peer-deps-external": "^2.2.2",
-    "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.31.2",
-    "rollup-plugin-visualizer": "^5.6.0",
+    "rollup-plugin-visualizer": "^5.12.0",
     "semantic-release": "^19.0.2",
     "semantic-release-export-data": "^1.0.1",
     "storybook-addon-designs": "^6.2.1",
@@ -173,12 +178,15 @@
     "styled-components": "^5.3.3",
     "stylelint": "^16.6.1",
     "stylelint-config-standard": "^36.0.0",
-    "ts-jest": "^27.1.4",
+    "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
     "ts-react-display-name": "^1.2.2",
-    "tslib": "^2.3.1",
-    "typescript": "^4.7.4",
-    "typescript-plugin-styled-components": "^2.0.0",
-    "yarn": "^1.22.22"
+    "tslib": "^2.6.2",
+    "typescript": "^5.5.2",
+    "typescript-plugin-styled-components": "^3.0.0",
+    "vite": "^5.2.8",
+    "vite-plugin-dts": "^3.8.1",
+    "vite-require": "^0.2.3"
   },
   "resolutions": {
     "trim": "1.0.1",
@@ -187,9 +195,10 @@
     "glob-parent": "5.1.2",
     "set-value": "4.1.0",
     "ansi-regex": "5.0.1",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "@emotion/react": "11.7.1",
+    "@storybook/react-docgen-typescript-plugin": "1.0.6--canary.9.cd77847.0",
     "postcss": "^8.4.38",
     "braces": "^3.0.3"
   },
