@@ -1,3 +1,3 @@
 export { default as Datatable } from './Datatable';
 export * from './Datatable.types';
-export { CustomColumnOptions } from './Table/Table.types';
+export type { CustomColumnOptions } from './Table/Table.types';
