@@ -1,2 +1,8 @@
 export { default as DatatableV2 } from './Datatable';
-export { DatatableColumnDef, DatatableOptions } from './Datatable.types';
+export type {
+  DatatableColumnDef,
+  DatatableOptions,
+  DatatableInstance,
+  DatatableRow,
+  DatatableRowAction,
+} from './Datatable.types';
