@@ -2,7 +2,7 @@ import React from 'react';
 
 import { DrawerSizes } from './Drawer.enums';
 
-export type DrawerSizes = typeof DrawerSizes[keyof typeof DrawerSizes];
+export type DrawerSizes = (typeof DrawerSizes)[keyof typeof DrawerSizes];
 
 export interface DrawerProps {
   children: React.ReactNode;
