@@ -9,7 +9,7 @@ import {
 } from '../_internal/BaseDateRangePicker/BaseDateRangePicker.types';
 import { Operators } from './Filters.enums';
 
-type OperatorTypes = typeof Operators[keyof typeof Operators];
+type OperatorTypes = (typeof Operators)[keyof typeof Operators];
 
 interface ComponentProps {
   options?: Option[];
@@ -66,7 +66,7 @@ export interface FiltersProps {
   isLoading?: boolean;
   isCancelEnabled?: boolean;
   isOperatorFieldEnabled?: boolean;
-  defaultOperator?: typeof Operators[keyof typeof Operators];
+  defaultOperator?: (typeof Operators)[keyof typeof Operators];
 }
 
 export const FieldPropTypes = PropTypes.exact({
