@@ -3,9 +3,10 @@ import { ElementType, ReactNode } from 'react';
 import { HeadingSizes, HeadingVariants } from './Heading.enums';
 import { SpacingProps } from '../../types/spacing.types';
 
-export type HeadingSizesTypes = typeof HeadingSizes[keyof typeof HeadingSizes];
+export type HeadingSizesTypes =
+  (typeof HeadingSizes)[keyof typeof HeadingSizes];
 export type HeadingVariantsTypes =
-  typeof HeadingVariants[keyof typeof HeadingVariants];
+  (typeof HeadingVariants)[keyof typeof HeadingVariants];
 
 export interface HeadingProps extends SpacingProps {
   size?: HeadingSizesTypes;
