@@ -5,8 +5,8 @@ import { Color } from '../../theme/colors.types';
 import { IconTypes, SSCIconNames } from '../../theme/icons/icons.enums';
 import { SpacingProps } from '../../types/spacing.types';
 
-export type SSCIcons = typeof SSCIconNames[keyof typeof SSCIconNames];
-export type Types = typeof IconTypes[keyof typeof IconTypes];
+export type SSCIcons = (typeof SSCIconNames)[keyof typeof SSCIconNames];
+export type Types = (typeof IconTypes)[keyof typeof IconTypes];
 // eslint-disable-next-line @typescript-eslint/ban-types
 export type IconNames = SSCIcons | (string & {});
 export type RegularIconTypes = Exclude<Types, 'fal'>;
