@@ -2,7 +2,7 @@ import React from 'react';
 
 import { ModalSizes } from './Modal.enums';
 
-export type ModalSizes = typeof ModalSizes[keyof typeof ModalSizes];
+export type ModalSizes = (typeof ModalSizes)[keyof typeof ModalSizes];
 
 export interface ModalProps {
   children: React.ReactNode;
