@@ -1,6 +1,6 @@
 export * as SignalEnums from './Signal.enums';
 export { default as Signal } from './Signal';
-export {
+export type {
   BreachRiskSignalKinds,
   BusinessImpactSignalKinds,
   IncidentLikelihoodSignalKinds,
