@@ -1,7 +1,8 @@
 import { TextSizes, TextVariants } from './Text.enums';
 
-export type TextSizesTypes = typeof TextSizes[keyof typeof TextSizes];
-export type TextVariantsTypes = typeof TextVariants[keyof typeof TextVariants];
+export type TextSizesTypes = (typeof TextSizes)[keyof typeof TextSizes];
+export type TextVariantsTypes =
+  (typeof TextVariants)[keyof typeof TextVariants];
 
 export interface TextProps {
   size?: TextSizesTypes;
