@@ -3,7 +3,7 @@ import type { To } from 'history';
 import { UserAvatarSizes } from './UserAvatar.enums';
 
 export type UserAvatarSizes =
-  typeof UserAvatarSizes[keyof typeof UserAvatarSizes];
+  (typeof UserAvatarSizes)[keyof typeof UserAvatarSizes];
 
 export type UserAvatarRootProps = {
   $isInverted?: boolean;
