@@ -1,3 +1,3 @@
 import { LinkColors } from './BaseLink.enums';
 
-export type Colors = typeof LinkColors[keyof typeof LinkColors];
+export type Colors = (typeof LinkColors)[keyof typeof LinkColors];
