@@ -2,7 +2,7 @@ import { ReactNode } from 'react';
 
 import { MessageVariants } from './Message.enums';
 
-export type Variants = typeof MessageVariants[keyof typeof MessageVariants];
+export type Variants = (typeof MessageVariants)[keyof typeof MessageVariants];
 
 export interface MessageProps {
   variant?: Variants;
