@@ -50,7 +50,7 @@ Select.propTypes = {
   maxPillLabelLength: PropTypes.number,
   maxVisibleItem: PropTypes.number,
   isMenuPositionRelative: PropTypes.bool,
-  components: PropTypes.shape(),
+  components: PropTypes.shape({}),
 };
 
 Select.defaultProps = {
