@@ -2,7 +2,7 @@ import React from 'react';
 
 import { SwitchSizes } from './Switch.enums';
 
-export type Sizes = typeof SwitchSizes[keyof typeof SwitchSizes];
+export type Sizes = (typeof SwitchSizes)[keyof typeof SwitchSizes];
 
 export interface SwitchLabelProps {
   isDisabled?: boolean;
