@@ -2,7 +2,7 @@ import { colors } from './colors';
 
 export type Color = {
   [K in keyof typeof colors]: `${K}.${Extract<
-    keyof typeof colors[K],
+    keyof (typeof colors)[K],
     string | number
   >}`;
 }[keyof typeof colors];
